Clarify replica shard allocation step in rolling restart procedure
Updates the restart documentation to specify that setting `cluster.routing.allocation.enable` to "primaries" disables **replica** shard allocation, not all shard allocation.

The phrase “disable shard allocation” seems to be a bit of a colloquial shorthand for developers and experienced users. Those familiar with Elasticsearch may implicitly understand it refers to replica shards during restarts. This PR makes that behavior explicit for newer users and reducing the risk of misconfiguration.