Clarify the steps of identifying the best ZK leader candidate
## Description

Previously it was documented as users / customer can find it thru logs.

Recently during a sync with @pfcoperez in an internal ticket - https://github.com/elastic/sdh-control-plane/issues/9169#issuecomment-2752931242 and https://github.com/elastic/support-tech-lead/issues/1554#issuecomment-2777954862, we decide to rewrite this part.

In detail, we'd recommend users / customers to only collect the essential information, and we (support) check the essential information is collected first, and engage with dev team to make further verification on this together.

Motivation is because if handled in a wrong way - either identify the best ZK leader candidate, or recover, this process can potentially corrupt users data permanently.

In specific,
- We hide the steps to identify the ZK leader into private section in KB:
  - Private view: https://support.elastic.dev/knowledge/view/fa410d1f
  - Public view: https://support.elastic.co/knowledge/fa410d1f
- We only guide users / customer to collect essential information, and let them reach out to support