Clarify the latest version of podman 5.2.2
## Description

Previously the latest version of podman 5.2.2 was 5.2.2-13. However, now the latest version is 5.2.2-15.

We have confirmed with that @matt-elastic in an [internal discussion here](https://elastic.slack.com/archives/CS5UH0E3T/p1748867788234029?thread_ts=1748577885.885639&cid=CS5UH0E3T) that it makes sense to make such change.

Thus, we need to update the public doc to reflect this.
Also, we want to avoid saying 5.2.2-15 and then later we need to update the public docs again, so we are using a more generic way to express it.