Adds a tip that explains how to install other 9.x versions
In case someone needs to install an older version of the Elasticsearch 9.x series, I'm adding a tip admonition to explain that they can replace the versions in the code examples with the version that they want. I'm using variables so that this guidance isn't hardcoded and doesn't require version bumps in the future.
Fixes #2234