Adding links to related pages
There are no details about the default ILM policy settings or index template options in the Rollover tutorial, so I'm adding them as these topics are relevant to expore. I'm also adding a link to the page about viewing the lifecycle of an existing index or data stream into the tutorial.
Fixes #1562