ILM docs: improve @custom component template usage/details (#3511)
In the ILM docs, the [Customize built-in
policies](https://www.elastic.co/docs/manage-data/lifecycle/index-lifecycle-management/tutorial-customize-built-in-policies#example-using-index-lifecycle-policy-apply-policy)
tutorial mentions `@custom` component templates with a "learn more" link
that points to the "Create or update component template" API docs. Not
super helpful, since those API docs don't mention the `@custom`
component templates.

Those templates are discussed in more detail in the Fleet docs, with
lengthy examples of using them. This PR adds a link to that content and
removes the API link.

I've also added a sentence at the top giving an example of creating a
custom policy for K8s logs, since we're told that those can be giant.

Closes: https://github.com/elastic/docs-content/issues/1557

---------

Co-authored-by: Edu González de la Herrán <25320357+eedugon@users.noreply.github.com>