Update prepare-to-upgrade.md to add ECH applies_to
The webpage [Upgrade on Elastic Cloud Hosted (ECH)](https://www.elastic.co/docs/deploy-manage/upgrade/deployment-or-cluster/upgrade-on-ech) is for ECH, but the first thing it says to do is go to `Prepare to upgrade` which has a `applies_to` for only Stack, but its linked from an `ECH` webpage. It would make more sense if the `Prepare to upgrade` also applies to `ECH`. I followed formatting compared to other webpages.