Update ECE troubleshooting docs
This is from an [internal discussion here](https://elastic.slack.com/archives/CKJ209XJL/p1764836058110659?thread_ts=1764835256.713879&cid=CKJ209XJL) that we want to make it more clear what users could do when they observe more than half of the directors (where zookeeper nodes running on) are lost, and thus caused the zookeeper quorum broken. 

In this situation, we have a doc talking about users can gather essential information and we can help them investigate, if there's still a possibility to recover the Zookeeper leader or not. (It depends on the status and this recovery is not guaranteed.)
https://www.elastic.co/docs/troubleshoot/deployments/cloud-enterprise/rebuilding-broken-zookeeper-quorum#ece_determine_the_zookeeper_leader