Fix up S3 repo logging docs
The docs about verbose logging from the AWS Java SDK only apply to the
EOL v1 SDK that we stopped using in 9.1. This commit adjusts the
documentation to reflect the new logger names from 9.1 onwards.

However, due to elastic/elasticsearch#137556 there was no way to enable
this logging at all in 9.1.0-9.1.8 and 9.2.0-9.2.2.